Update ShowPreviousParseTest, RetrievePreviousCommand, ShowPreviousParse
